Hacking on PWM driver for RGB, doesn't work yet.
ARM doesn't support split sync with the default bitbang driver for
WS2812 control, so PWM has to be used. But the docs aren't very complete
on how to properly enable the right combination of DMA channels, PWM
timers, etc. and I haven't found the right combination of settings yet.
That leaves driving the LEDs with the bitbang driver from a single side
and using one of the extra conductors in the interconnect cable.